feat(web-server): Validate payload for IP filters

+++ backend/web-server/controllers/ip-filter.controller.ts
@@ -1,5 +1,6 @@
-import { KoaContext } from '../koa';
+import { KoaContext, KoaError } from '../koa';
 import { IpFilterCommandDTO, IpFilterDTO } from '../../../shared/model/ip-filter.model';
+import oibusSchema from '../../engine/oibus-validation-schema';
 
 const getIpFilters = async (ctx: KoaContext<void, Array<IpFilterDTO>>) => {
   const ipFilters = ctx.app.repositoryService.ipFilterRepository.getIpFilters();
@@ -8,32 +9,41 @@ const getIpFilters = async (ctx: KoaContext<void, Array<IpFilterDTO>>) => {
 
 const getIpFilter = async (ctx: KoaContext<void, IpFilterDTO>) => {
   const ipFilter = ctx.app.repositoryService.ipFilterRepository.getIpFilter(ctx.params.id);
-  ctx.ok(ipFilter);
+  if (ipFilter) {
+    ctx.ok(ipFilter);
+  } else {
+    ctx.notFound();
+  }
 };
 
 const createIpFilter = async (ctx: KoaContext<IpFilterCommandDTO, void>) => {
-  const command: IpFilterCommandDTO | undefined = ctx.request.body;
-  if (command) {
-    const ipFilter = ctx.app.repositoryService.ipFilterRepository.createIpFilter(command);
+  try {
+    const ipFilterCommandDTO: IpFilterCommandDTO = await oibusSchema.ipFilterSchema.validateAsync(ctx.request.body);
+    const ipFilter = ctx.app.repositoryService.ipFilterRepository.createIpFilter(ipFilterCommandDTO);
     ctx.created(ipFilter);
-  } else {
-    ctx.badRequest();
+  } catch (error: any) {
+    ctx.badRequest(error.message);
   }
 };
 
 const updateIpFilter = async (ctx: KoaContext<IpFilterCommandDTO, void>) => {
-  const command: IpFilterCommandDTO | undefined = ctx.request.body;
-  if (command) {
-    ctx.app.repositoryService.ipFilterRepository.updateIpFilter(ctx.params.id, command);
+  try {
+    const ipFilterCommandDTO: IpFilterCommandDTO = await oibusSchema.ipFilterSchema.validateAsync(ctx.request.body);
+    ctx.app.repositoryService.ipFilterRepository.updateIpFilter(ctx.params.id, ipFilterCommandDTO);
     ctx.noContent();
-  } else {
-    ctx.badRequest();
+  } catch (error: any) {
+    ctx.badRequest(error.message);
   }
 };
 
 const deleteIpFilter = async (ctx: KoaContext<void, void>) => {
-  ctx.app.repositoryService.ipFilterRepository.deleteIpFilter(ctx.params.id);
-  ctx.noContent();
+  const ipFilter = ctx.app.repositoryService.ipFilterRepository.getIpFilter(ctx.params.id);
+  if (ipFilter) {
+    ctx.app.repositoryService.ipFilterRepository.deleteIpFilter(ctx.params.id);
+    ctx.noContent();
+  } else {
+    ctx.notFound();
+  }
 };
 
 export default {
